refactor: RacingCars 객체로 경주 역할 분리
